requirements.txt
collision==1.2.2
numba==0.53.1
numpy==1.20.2
Pillow==9.0.1
pyglet==1.5.17
screeninfo==0.8
termcolor==1.1.0
py-cpuinfo==8.0.0
psutil==5.9.1